requirements.txt
flask
pymongo
python-dotenv
flask-login
Werkzeug